format: reorganizes dir and controller, then adds sticky footer
