+++ src/App.js
@@ -3,9 +3,9 @@ import React, { Component } from 'react';
 import { Switch, Route } from 'react-router-dom';
 
 import About from './containers/About';
-import FormContainer from './containers/FormContainer';
+import Form from './containers/Form';
 import Home from './containers/Home';
-import Thanks from './containers/Thanks';
+import Thanks from './components/Thanks';
 
 import RentalDataStore from './rentalDataStore';
 
@@ -22,9 +22,8 @@ class App extends Component {
           <Switch>
             <Route exact path='/' component={Home} />
             <Route exact path='/about' component={About} />
-            <Route exact path='/form' component={FormContainer} />
+            <Route exact path='/form' component={Form} />
             <Route exact path='/thanks' component={Thanks} />
-
           </Switch>
         </Provider>
       </div>
