+++ src/components/Footer.js
@@ -1,5 +1,25 @@
 import React, { Component } from 'react';
 
+let sticky = {
+  backgroundColor: "#F8F8F8",
+  borderTop: "1px solid #E7E7E7",
+  textAlign: "center",
+  padding: "20px",
+  position: "fixed",
+  left: "0",
+  bottom: "0",
+  height: "60px",
+  width: "100%",
+}
+
+let phantom = {
+  display: 'block',
+  padding: '20px',
+  height: '60px',
+  width: '100%',
+}
+
+
 class Footer extends Component {
   constructor(props) {
     super(props)
@@ -9,6 +29,7 @@ class Footer extends Component {
   }
   render() {
     return (
+      <div style={{...sticky, ...phantom}}>
       <footer>
         <ul className="site-link">
           <li>
@@ -16,6 +37,7 @@ class Footer extends Component {
           </li>
         </ul>
       </footer>
+      </div>
     );
   }
 }
