+++ src/styles/style.css
@@ -38,3 +38,23 @@ hr {
 Col {
   padding: 20px;
 }
+
+.sticky {
+  background-color: "#F8F8F8";
+  border-top: "1px solid #E7E7E7";
+  text-align: "center";
+  padding: "20px";
+  position: "fixed";
+  left: "0";
+  bottom: "0";
+  height: "60px";
+  width: "100%";
+}
+
+.phantom {
+    display: 'block';
+    padding: '20px';
+    height: '60px';
+    width: '100%';
+  
+}
